# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "CreateMemoryRequest": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "memory_type": {
            "enum": ["fact", "preference", "project", "reference", "episodic"],
            "title": "Memory Type",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Derived from title when omitted",
            "title": "Slug"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["title", "content", "memory_type"],
        "title": "CreateMemoryRequest",
        "type": "object"
      },
      "EditMemoryRequest": {
        "description": "Sparse edit; omitted fields stay unchanged.",
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "memory_type": {
            "anyOf": [
              {
                "enum": ["fact", "preference", "project", "reference", "episodic"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Type"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": ["expected_version"],
        "title": "EditMemoryRequest",
        "type": "object"
      },
...
      "LinkView": {
        "properties": {
          "relation": {
            "enum": ["relates_to", "supersedes", "derived_from", "contradicts"],
            "title": "Relation",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": ["target_id", "relation"],
        "title": "LinkView",
        "type": "object"
      },
      "MemoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items"],
        "title": "MemoryListResponse",
        "type": "object"
      },
      "MemoryResponse": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkView"
            },
            "title": "Links",
            "type": "array"
          },
          "memory_type": {
            "enum": ["fact", "preference", "project", "reference", "episodic"],
            "title": "Memory Type",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "description": "Optimistic concurrency token; echo it in edits",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "content",
          "memory_type",
          "tags",
          "links",
          "archived",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "MemoryResponse",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details \u2014 the error shape of every non-2xx response.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["title", "status"],
        "title": "Problem",
        "type": "object"
      },
...
      "TimelineEntryResponse": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "stream_seq": {
            "title": "Stream Seq",
            "type": "integer"
          }
        },
        "required": ["event_id", "event_type", "occurred_at", "actor", "stream_seq"],
        "title": "TimelineEntryResponse",
        "type": "object"
      },
      "TimelineResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "memory_id": {
            "format": "uuid",
            "title": "Memory Id",
            "type": "string"
          }
        },
        "required": ["memory_id", "entries"],
        "title": "TimelineResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Git-native, event-sourced, user-owned memory for AI assistants.",
    "title": "engram API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/rebuild": {
      "post": {
        "description": "Replay the event log through every projection from global_seq 0.",
        "operationId": "rebuild_projections_admin_rebuild_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Rebuild Projections Admin Rebuild Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Rebuild Projections",
        "tags": ["system"]
      }
    },
...
    "/api/v1/memories": {
      "get": {
        "description": "Cursor-paginated listing of current memory state.",
        "operationId": "list_memories_api_v1_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "memory_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": ["fact", "preference", "project", "reference", "episodic"],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Memory Type"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "List Memories",
        "tags": ["memories"]
      },
      "post": {
        "description": "Create a memory; returns its immutable id and initial state.",
        "operationId": "create_memory_api_v1_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Create Memory",
        "tags": ["memories"]
      }
    },
    "/api/v1/memories/{memory_id}": {
      "delete": {
        "description": "Tombstone the memory (its history remains in the event log).",
        "operationId": "delete_memory_api_v1_memories__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Delete Memory",
        "tags": ["memories"]
      },
      "get": {
        "description": "Current state of one memory.",
        "operationId": "get_memory_api_v1_memories__memory_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Get Memory",
        "tags": ["memories"]
      },
      "patch": {
        "description": "Sparse edit. 409 when ``expected_version`` is stale.",
        "operationId": "edit_memory_api_v1_memories__memory_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Edit Memory",
        "tags": ["memories"]
      }
    },
    "/api/v1/memories/{memory_id}/timeline": {
      "get": {
        "description": "Full event history of one memory, oldest first.",
        "operationId": "memory_timeline_api_v1_memories__memory_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Memory Timeline",
        "tags": ["memories"]
      }
    },
    "/api/v1/memories/{memory_id}/undo": {
      "post": {
        "description": "Append the compensating event for the most recent change.",
        "operationId": "undo_last_change_api_v1_memories__memory_id__undo_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Undo Last Change",
        "tags": ["memories"]
      }
    },
...
  }
}